+++ Src/Private/Get-AbrADSite.ps1
@@ -30,23 +30,25 @@ function Get-AbrADSite {
                     Paragraph 'Replication is the process of transferring and updating Active Directory objects between domain controllers in the Active Directory domain and forest.'
                     BlankLine
                     Paragraph "The following section details Active Directory replication and its relationships."
-                    try {
+                    if ($Options.EnableDiagrams) {
                         try {
-                            $Graph = New-ADDiagram -Target $System -Format base64 -Direction top-to-bottom -DiagramType SitesInventory -PSSessionObject $TempPssSession
-                        } catch {
-                            Write-PScriboMessage -IsWarning "Site Inventory Diagram Graph: $($_.Exception.Message)"
-                        }
-
-                        if ($Graph) {
-                            If ((Get-DiaImagePercent -GraphObj $Graph).Width -gt 1500) { $ImagePrty = 20 } else { $ImagePrty = 50 }
-                            Section -Style Heading4 "Site Inventory Diagram." {
-                                Image -Base64 $Graph -Text "Site Inventory Diagram" -Percent $ImagePrty -Align Center
-                                Paragraph "Image preview: Opens the image in a new tab to view it at full resolution." -Tabs 2
+                            try {
+                                $Graph = Get-AbrDiagrammer -DiagramType "SitesInventory" -DiagramOutput base64 -PSSessionObject $TempPssSession
+                            } catch {
+                                Write-PScriboMessage -IsWarning "Site Inventory Diagram Graph: $($_.Exception.Message)"
                             }
-                            BlankLine -Count 2
+
+                            if ($Graph) {
+                                If ((Get-DiaImagePercent -GraphObj $Graph).Width -gt 1500) { $ImagePrty = 20 } else { $ImagePrty = 50 }
+                                Section -Style Heading4 "Site Inventory Diagram." {
+                                    Image -Base64 $Graph -Text "Site Inventory Diagram" -Percent $ImagePrty -Align Center
+                                    Paragraph "Image preview: Opens the image in a new tab to view it at full resolution." -Tabs 2
+                                }
+                                BlankLine -Count 2
+                            }
+                        } catch {
+                            Write-PScriboMessage -IsWarning "Site Inventory Diagram Section: $($_.Exception.Message)"
                         }
-                    } catch {
-                        Write-PScriboMessage -IsWarning "Site Inventory Diagram Section: $($_.Exception.Message)"
                     }
                     Section -Style Heading4 'Sites' {
                         $OutObj = @()
@@ -323,23 +325,25 @@ function Get-AbrADSite {
                     } catch {
                         Write-PScriboMessage -IsWarning "$($_.Exception.Message) (Site Subnets)"
                     }
-                    try {
+                    if ($Options.EnableDiagrams) {
                         try {
-                            $Graph = New-ADDiagram -Target $System -Format base64 -Direction top-to-bottom -DiagramType Sites -PSSessionObject $TempPssSession
-                        } catch {
-                            Write-PScriboMessage -IsWarning "Site Topology Diagram Graph: $($_.Exception.Message)"
-                        }
-
-                        if ($Graph) {
-                            If ((Get-DiaImagePercent -GraphObj $Graph).Width -gt 1500) { $ImagePrty = 10 } else { $ImagePrty = 50 }
-                            Section -Style Heading4 "Site Topology Diagram." {
-                                Image -Base64 $Graph -Text "Site Topology Diagram" -Percent $ImagePrty -Align Center
-                                Paragraph "Image preview: Opens the image in a new tab to view it at full resolution." -Tabs 2
+                            try {
+                                $Graph = Get-AbrDiagrammer -DiagramType "Sites" -DiagramOutput base64 -PSSessionObject $TempPssSession
+                            } catch {
+                                Write-PScriboMessage -IsWarning "Site Topology Diagram Graph: $($_.Exception.Message)"
                             }
-                            BlankLine -Count 2
+
+                            if ($Graph) {
+                                If ((Get-DiaImagePercent -GraphObj $Graph).Width -gt 1500) { $ImagePrty = 10 } else { $ImagePrty = 50 }
+                                Section -Style Heading4 "Site Topology Diagram." {
+                                    Image -Base64 $Graph -Text "Site Topology Diagram" -Percent $ImagePrty -Align Center
+                                    Paragraph "Image preview: Opens the image in a new tab to view it at full resolution." -Tabs 2
+                                }
+                                BlankLine -Count 2
+                            }
+                        } catch {
+                            Write-PScriboMessage -IsWarning "Site Topology Diagram Section: $($_.Exception.Message)"
                         }
-                    } catch {
-                        Write-PScriboMessage -IsWarning "Site Topology Diagram Section: $($_.Exception.Message)"
                     }
                     try {
                         $DomainDN = Invoke-Command -Session $TempPssSession { (Get-ADDomain -Identity (Get-ADForest | Select-Object -ExpandProperty RootDomain )).DistinguishedName }