+++ Src/Private/Get-AbrADTrust.ps1
@@ -108,23 +108,25 @@ function Get-AbrADTrust {
                                 }
                                 $TrustInfo | Table @TableParams
                             }
-                            try {
+                            if ($Options.EnableDiagrams) {
                                 try {
-                                    $Graph = New-ADDiagram -Target $DC -Credential $Credential -Format base64 -Direction top-to-bottom -DiagramType Trusts
-                                } catch {
-                                    Write-PScriboMessage -IsWarning "Domain and Trusts Diagram Graph: $($_.Exception.Message)"
-                                }
-
-                                if ($Graph) {
-                                    If ((Get-DiaImagePercent -GraphObj $Graph).Width -gt 1500) { $ImagePrty = 10 } else { $ImagePrty = 50 }
-                                    Section -Style Heading3 "Domain and Trusts Diagram." {
-                                        Image -Base64 $Graph -Text "Domain and Trusts Diagram" -Percent $ImagePrty -Align Center
-                                        Paragraph "Image preview: Opens the image in a new tab to view it at full resolution." -Tabs 2
+                                    try {
+                                        $Graph = Get-AbrDiagrammer -DiagramType "Trusts" -DiagramOutput base64 -PSSessionObject $TempPssSession
+                                    } catch {
+                                        Write-PScriboMessage -IsWarning "Domain and Trusts Diagram Graph: $($_.Exception.Message)"
                                     }
-                                    BlankLine -Count 2
+
+                                    if ($Graph) {
+                                        If ((Get-DiaImagePercent -GraphObj $Graph).Width -gt 1500) { $ImagePrty = 10 } else { $ImagePrty = 50 }
+                                        Section -Style Heading3 "Domain and Trusts Diagram." {
+                                            Image -Base64 $Graph -Text "Domain and Trusts Diagram" -Percent $ImagePrty -Align Center
+                                            Paragraph "Image preview: Opens the image in a new tab to view it at full resolution." -Tabs 2
+                                        }
+                                        BlankLine -Count 2
+                                    }
+                                } catch {
+                                    Write-PScriboMessage -IsWarning "Domain and Trusts Diagram Section: $($_.Exception.Message)"
                                 }
-                            } catch {
-                                Write-PScriboMessage -IsWarning "Domain and Trusts Diagram Section: $($_.Exception.Message)"
                             }
                         }
                     } else {